New home page and shift question page to its own folder
