+++ apps/frontend/src/app/question/styles.scss
@@ -1,0 +1,64 @@
+.layout {
+  background: white;
+}
+
+.content {
+  background: white;
+}
+
+.content-card {
+  margin: 4rem 8rem;
+  padding: 2rem 4rem;
+  background-color: white;
+  min-height: 100vh;
+  border-radius: 30px;
+  box-shadow: 0px 10px 60px rgba(226, 236, 249, 0.5);
+}
+
+.content-row-1 {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.content-title {
+  //   font-family: "Poppins";
+  //   font-style: normal;
+  font-weight: 600;
+  font-size: 22px;
+  line-height: 33px;
+  letter-spacing: -0.01em;
+  color: #000000;
+}
+
+.content-filter {
+  margin: 1.5rem 0;
+}
+
+.edit-button {
+  margin-right: 0.5rem;
+}
+
+.filter-button {
+  margin-left: 8px;
+  box-shadow: 0 2px 0 rgba(0, 0, 0, 0.02) !important;
+}
+
+.clear-button {
+  width: 100%;
+}
+
+.categories-multi-select,
+.difficulty-select,
+.order-select {
+  width: 100%;
+}
+
+.create-title,
+.new-problem-categories-multi-select,
+.new-problem-difficulty-select,
+.create-description,
+.create-problem-id {
+  width: 100%;
+  margin: 5px;
+}
